+++ png.h
@@ -8,19 +8,19 @@
 
 class Png : public AspngSurface {
 public:
-    Png(size_t, size_t);
+    Png(int32_t, int32_t);
     static Png *read(std::string);
     void write(std::string);
-    size_t get_width();
-    size_t get_height();
-    Rgb get_pixel(size_t, size_t);
-    void set_pixel(size_t, size_t, Rgb);
+    int32_t get_width();
+    int32_t get_height();
+    Rgb get_pixel(int32_t, int32_t);
+    void set_pixel(int32_t, int32_t, Rgb);
 
 private:
     Png();
 
-    size_t width;
-    size_t height;
+    int32_t width;
+    int32_t height;
     png_bytep *rows;
 };
